GH-744: started converting gas_price units from gwei to wei

+++ node/src/accountant/scanners/mid_scan_msg_handling/payable_scanner/agent_null.rs
@@ -27,7 +27,7 @@ impl BlockchainAgent for BlockchainAgentNull {
         }
     }
 
-    fn agreed_fee_per_computation_unit(&self) -> u64 {
+    fn agreed_fee_per_computation_unit(&self) -> u128 {
         self.log_function_call("agreed_fee_per_computation_unit()");
         0
     }
