+++ node/src/accountant/scanners/mid_scan_msg_handling/payable_scanner/blockchain_agent.rs
@@ -24,7 +24,7 @@ use web3::types::U256;
 pub trait BlockchainAgent: Send {
     fn estimated_transaction_fee_total(&self, number_of_transactions: usize) -> u128;
     fn consuming_wallet_balances(&self) -> ConsumingWalletBalances;
-    fn agreed_fee_per_computation_unit(&self) -> u64;
+    fn agreed_fee_per_computation_unit(&self) -> u128;
     fn consuming_wallet(&self) -> &Wallet;
     fn pending_transaction_id(&self) -> U256;
